feat: Adds first, simple app `forty-two.pnt` (WIP)
This adds a trivial, simple app that allows for:

1. Providing the answer to `6 * 7` as a decision variable.
2. Updating an address if the constraint passes.

The goal for this app is to be a simple, minimal PoC that demonstrates
how to build, sign, deploy and solve a simple pint intent set with
a small bash script.

The bash script can be run by first using `nix shell` to bring all
required tools into scope, and then running the script.

To-Do
-----

For the most part, it hasn't been *too* tricky to build using `pintc`,
manipulate the JSON into the necessary layout using `jq`, and interact
with the server via `curl`. However there have been a few large
hinderances that would be nice to solve before landing:

- [ ] Create a minimal tool for deserializing intent sets and solutions,
      signing them, and re-serializing the `Signed` form. Currently,
      there's no way to achieve this without writing a custom Rust
      program as we must ensure the hashing and signing methods used are
      exact.
- [ ] Add solution and script for submitting and causing state update.
- [ ] Add CI app testing process that tests the full end-to-end
      intent-set deployment and solution submission and checks the
      resulting state of the server.

Follow-up
---------

- [ ] Serialize the following as hex strings. Currently, they are all
      serialized to JSON as massive decimal byte lists. A more compact
      string representation might help to keep the JSON more compact
      until we work out how we want to send blobs alongside the json?
      - `StateReadBytecode`
      - `ConstraintBytecode`
      - `Signature` `data` field
      - `ContentAddress`
      Perhaps we can use the [`is_human_readable`](https://docs.rs/serde/latest/serde/trait.Serializer.html#method.is_human_readable)
      check to only serialize/deserialize via string in the case that the
      format is intended for human readability?
- [ ] Land tracing in server upstream: Currently very difficult to track
      what is going on.